[#59505526] enable schema checking in Vcloud::Launch
If a schema is supplied to config_loader, then we will check against
it, log errors, and raise an exception if validation fails.

Added tests in config_loader_spec.rb to check validation, and also to
check validation failure.

The tests for config_loader are very specific to Vcloud::Launch though,
it might be worth splitting these out into their respective modules later.
This could happen when we switch Vcloud::NetLaunch to use validation too.