Add test that confirms no update
The majority of the logic around when update is called will be in unit tests. However it is worth adding one integration test to confirm that no update takes place.

The way this is tested is by getting the id of the last task that was run, the point being that new task = new id, so same id = no task was run, i.e. no update. This is slightly oblique - a better way to do this would be add functionality to Core to get the last task and then from that get the timestamp of the last update, but that is a lot of work for a test.

To clarify, I've added a comment, and an example of where they are different in a previous test.