Add HTTPS-based LoadBalancerService generator test
The load_balancer_https-output.yaml file was generated from a
live load balancer EdgeGateway, configured via the UI.

The load_balancer_https-input.yaml file was created manually to match.

Tests for this are currently failing, since there are assumptions in the
generator that are not marrying to the actual API. These will be fixed in
subsequent commits.