Clarify the update
Ideally everything in proposed_config takes place without reference to the EdgeGateway as we are passing in the local and remote config for the Differ. However, there is an outstanding issue with the NatService which makes some unnecessary calls which should be dependency injected. There is a story for this: https://www.pivotaltracker.com/story/show/65619490