Remove use of and_call_original
As per NatService specific tests, remove need to have rSpec and_call_original
call in the test path - this is risky in an integration test.

This also aligns the method that we are using to test how many times the
edgeGateway has been updated, so is ripe for DRYing out.