Fix mock data in EdgeGatewayConfiguration test
The test double of Core::EdgeGateway.vcloud_gateway_interface_by_id
is hardcoded to return a particular Interface section.

Make the inputs for tests match this double, otherwise the tests don't
really make sense.

This limitation with the mocking is addressed when we move to dependency
injection for the generators.