Convert NatService to use dependency injection
Rather than getting NatService generator to call out to
the EdgeGateway for each interface, utilise the
Core::EdgeGateway.interfaces method to supply an array
of EdgeGatewayInterface objects to it, which can then be
used to look up the needed network details.

Similarly, this interfaces list is supplied to the
EdgeGatewayConfiguration class by EdgeGatewayServices, ready
for use in other generators (specifically, LoadBalancerService)

Committing this change with one failing unit test in the
NatService generator - the test that checks that we only call
the EdgeGateway API once. This is no longer relevant, and will be
removed in the next commit.

Integration tests are passing.