Update default healthcheck URI to be '/'
A healthcheck URI of '' (which technically is the defualt if no
Uri is provided) results in a healthcheck request of 'OPTIONS *'.

This is rejected with a 400 by Nginx and likely other servers, and
so is not really a sensible default - vCloud Director considers the 400
a bad request, and marks the member node as down.

Instead, use '/' as the default healthcheck Uri, as this then maps
to a 'GET /' call, generally returning ok by most sites.

NB: This vCloud Director documentation suggests that the default is /
    regardless. This at least makes our tool match the docs.