Add FirewallConfigurationDiffer class to ignore Id
(as per the NatConfigurationDiffer class in daebc340)

When new firewall_rules are inserted, the resulting generator output
dutifully increments the Id value of FirewallRules that follow.

This creates a problem when comparing with ConfigurationDiffer,
since the rules after our insertion have not changed, bar their
Id value increasing in relation to the new rules.

This commit adds a FirewallConfigurationDiffer class which strips out
these Id values before comparison.