Do inheritance in a clearer way
There's no need for code to be repeated in every subclass. The only thing that really needs to change is what the unused field is.

The difference between this and what was there before is that the unused field is stripped from both local and remote whereas in the case of the Load Balancer, it won't be in the local. This doesn't matter though, as there is a test for whether it's present before attemptint to remove it.