I've added a lot of tests here
So coverage is higher. I'm not entirely convinced by this, because the tests were all in one area - it's clearly not testing lines of code or paths through the code. However, this is the tool we are using for coverage at the moment so might as well make use of it.