Fix assignment in condition lint warnings
This is confusing for the small amount of characters it saves. Most of these
can refer to the original variable instead.

The exception is `input_pool_service_port[:health_check]` which referred to
the new var twice and the old var once within the condition. I've replaced
the ternary condition with `.fetch(key, default).to_s` to do the same thing
and preserve line length.