Add explicit namespace to Query and QueryRunner
Add the explicit namespacing to Query and QueryRunner in line with the
changes introduced in vcloud-core version 0.2.0, in which these classes
were moved under the Vcloud::Core namespace.

Also bumps the dependency on vcloud-core in the Gemspec so that we use
version 0.2.0, and bumps vcloud-edge_gateway to version 0.3.0.

Prefer a minor version bump for vcloud-edge_gateway (as opposed to a
patch version bump) to make it clearer that the changes in vcloud-core
broke backwards compatibility and that it's preferable to upgrade.

Depends on alphagov/vcloud-core#33.