Make template_vars a named CLI argument
We have a story in the icebox about providing multiple input files to
`vcloud-launch` which, if implemented in Vcloud::Core::ConfigLoader, could
also be made available here.

This implies providing multiple file arguments on the command line which
will make it very hard to distinguish our presently-unnamed argument for
template_vars from the rest of the input files.

So preemptively make this a named argument and update the documentation
accordingly.

PS: The `if` condition isn't stritly neccessary because the hash will return
nil (which matches the arg default) if the option hasn't been provided. But
I think it's better to be explicit.