Rename EdgeGatewayServices->EdgeGateway::Configure
So that the class is namespaced under `Vcloud::EdgeGateway` like everything
else. I've renamed it generically to `Configure`, for want of a better name,
because:

- It matches the binary/class names in the other tools.
- It's "action" is to configure a VSE.

There is some overlap with `Vcloud::EdgeGateway::EdgeGatewayConfiguration`
which I've raised [#71220390] to investigate.

I've removed references to the class name in the test "context" blocks
because they are hierarchical and the name comes from the outer "describe"
block.