Be explicit about service port being tested
In the load balancer spec tests, we make assertions against the
`:ServicePort` array in the generated configuration.

In order to match the appropriate service port to the load balancer
service protocol being tested, we have to specify the array key by its
index.

To make it more obvious exactly what service port we are targetting,
add a new variable for the service port we are targeting and refer to
it in our assertions.

This does not change the functionality of the tests but makes them more
readable.