Restrict simplecov to unit tests
The addition of Vcloud::EdgeGateway::Cli, which was covered by unit tests,
caused the integration tests to fail on low coverage because they are run
from a separate rake task.

Use the COVERAGE environment variable, already set by the Rakefile for unit
tests, to restrict coverage checking. This aligns it with how vcloud-core
works.