Default colour to off when redirecting STDOUT
Disable coloured output when STDOUT is redirected to something that isn't
associated with a TTY. It can be overridden with `--colour`.

This copies the default behaviour of utilities like `ls` and `grep` so that
the output isn't littered with ANSI codes when you're redirecting the output
to a file or piping to another utility for further processing.