Reverse args for calls to *Differ
Reverse the arguments in calls to:

- FirewallConfigurationDiffer()
- NatConfigurationDiffer()
- LoadBalancerConfigurationDiffer()

So that they are remote->local instead of local->remote. This will make the
diffs readable when we come to expose them. By being the "wrong" way round
it looked like we were removing rules when we were infact adding them.

This has no functional effect on the current behaviour of
EdgeGatewayConfiguration because a change, in either direction, is worthy of
triggering a new config for that service.