Refactor diff tests for Cli
Use shared_examples to DRY up the tests for Cli diff output. I'm about to
add another two tests like this and don't want the intended behaviours to
get lost in a sea of code. I'm `it_behaves_like` on the outter call so that
we get a meaningful name in the documentation and `include_example` on the
inner call which is hidden from the output. It looks as follows when using
`rspec -fd`:
```
  diff output
    when diff is empty
      when given config (colour doesn't matter)
        behaves like diff with stdout contents
          should output diff to stdout
    when diff contains two services with three types of changes
      when colour argument is not specified
        behaves like diff with colour output
          should output diff to stdout
      when given --no-colour
        behaves like diff without colour output
          should output diff to stdout
```