CLI output diffs with optional colour
Make the CLI tool output a diff of the changes between the remote config
(VSE) and the local config (file) that you're applying. This can be used to
determine what, if anything, changed during the run.

This is always returned to STDOUT. If you're not interested in it then it
could be redirected to `/dev/null`. The colour defaults to on in order to
make the additions and removals more human readable, but can be turned off
with `--no-colour`.

It's important to note that the output format is what's produced by
Vcloud::EdgeGateway::ConfigurationGenerator::* and NOT the config input
format, which means that it's much more verbose and doesn't contain exactly
the same key names. Unfortunately it's too difficult at this stage to
convert it back to the original input format.

I did consider rendering the output to YAML but there's a risk that people
might confuse it for the config format. I also considered pretty-printing
it, but all those nested hashes and newlines are going to produce a very
lengthy output. I think it is readable in it's current form though.

I chose to implement the render in EdgeGateway::Cli instead of
EdgeGateway::Configure or EdgeGateway::EdgeGatewayConfiguration because the
inner classes shouldn't need to have any knowledge about colour output and
it allows other consumers of those classes (treating it as a library) to
render their own diffs how they wish.