Expose diffs from EdgeGatewayConfiguration
Create a new class var/method which stores the diffs returned by the
service-specific differ classes into a hash which is keyed by service name.
The hash will be empty (no keys) when the local and remote configs are the
same.

This will be consumed by either EdgeGateway::Configure or EdgeGateway::Cli
in order to return the diffs alongside the services that need reconfiguring.