Return config diff from Configure#update
So that users of this class can inspect and present the diff for themselves.

This will primarily be for Vcloud::EdgeGateway::Cli, but may also be useful
for anyone wanting to use this gem as a library.

Shallow integration tests - check that it has the right keys and at least
one diff item for each - because we don't have any unit tests for Configure
and we need to know that this is working end-to-end. Especially because we
may not integration test EdgeGateway::Cli as the end consumer of this.

I've piggybacked onto the existing tests, some of which relate to "should
only make one call" or "should one create one task", because this is the
only place we have access to the diff (not the subsequent it block) and it's
too expensive (runtime) to duplicate these tests just to check the diff.
When discussed with Anna, we came conclusion that these weren't stretching
the existing descriptions too much.

The integration tests for the service-specific *Differ classes have been
removed because these new tests perform the same purpose more
comprehensively and without reaching outside EdgeGateway::Configure.