Amend help message for --[no-]colour
As noted by Matt, the word "on" doesn't quite match the values of "enabled"
or "disabled". Re-use the word "enabled" instead.

Also reverse the order of "enable/disable" to indicate that it would
normally be used to disable colour and to match the order of the arg spec
`--[no-]-colour`.